move fetchSharedId() & storeSharedId() to SharedId
